python/requirements.txt
setuptools
python-ldap
mysql-connector-python